range:  Fix the situation cannot drag to max value (Close #1556)
