+++ src/components/range/lib/closest.js
