+++ src/components/range/lib/delegate.js
