+++ src/components/range/lib/emitter.js
