+++ src/components/range/lib/event.js
