+++ src/components/range/lib/events.js
