+++ src/components/range/lib/mouse.js
