+++ src/components/range/lib/query.js
