+++ src/components/range/metas.yml
@@ -59,6 +59,11 @@ events:
     en: emits when value is changed
     zh-CN: 绑定值变化时触发事件
 changes:
+  next:
+    en:
+      - '[fix] Fix the situation cannot drag to max value #1556'
+    zh-CN:
+      - '[fix] 修复一些情况下无法拖到最大值的问题 #1556'
   v2.2.2:
     en:
       - '[feature] Support dynamically setting step #1394'
