--- src/components/range/range/lib/lib/indexof.js
@@ -1,7 +1,0 @@
-module.exports = function (arr, obj) {
-  if (arr.indexOf) return arr.indexOf(obj)
-  for (var i = 0; i < arr.length; ++i) {
-    if (arr[i] === obj) return i
-  }
-  return -1
-}
